% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLstudio.R
\name{shiny_package}
\alias{shiny_package}
\title{Lunch the ML Studio app}
\usage{
shiny_package(launch.browser = TRUE, seed = 1234)
}
\arguments{
\item{launch.browser}{A Boolean variable, if true will lunch the app on a web browser (default is “TRUE”)}
}
\description{
Lunch the ML Studio app
}
\examples{
MLstudio(launch.browser = TRUE)
}
